search: move query, search-page, run-action and unified-bar into panel-plugin/search/
