+++ panel-plugin/meson.build
@@ -24,14 +24,14 @@ plugin_sources = [
   'places/places-page.cpp',
   'plugin.cpp',
   'profile.cpp',
-  'query.cpp',
+  'search/query.cpp',
   'recent-page.cpp',
   'register-plugin.c',
   'resizer.cpp',
-  'run-action.cpp',
-  'search-action.cpp',
+  'search/run-action.cpp',
+  'search/search-action.cpp',
   'config/usage-stats.cpp',
-  'search-page.cpp',
+  'search/search-page.cpp',
   'settings.cpp',
   'settings-bindings.cpp',
   'settings-defaults.cpp',
