+++ panel-plugin/search/query.cpp
