+++ panel-plugin/search/query.h
