+++ panel-plugin/search/run-action.cpp
