+++ panel-plugin/search/run-action.h
