+++ panel-plugin/search/search-action.cpp
