+++ panel-plugin/search/search-action.h
