+++ panel-plugin/search/search-page.cpp
