+++ panel-plugin/search/search-page.h
