+++ panel-plugin/search/unified-bar.h
